Move NamedCollectionsFactory to a separate file
