+++ src/Common/NamedCollections/NamedCollectionUtils.cpp
@@ -16,6 +16,7 @@
 #include <Interpreters/Context.h>
 #include <Common/NamedCollections/NamedCollections.h>
 #include <Common/NamedCollections/NamedCollectionConfiguration.h>
+#include <Common/NamedCollections/NamedCollectionsFactory.h>
 
 #include <filesystem>
